refactor(block): decouple internal/external handling from edge executor
- Introduce block-side adapters:
  - manifestPoller: shared external JSONL/text polling and reconstruction
  - blockUpdateApplier: unify BLOCK_UPDATED handling and normalize content
- Update runSingleEdgeNodeExecutor/runAllNodesExecutor to delegate to adapters
- Result: editors always receive string content; edge no longer writes raw objects
